+++ src/snowcli/cli/snowpark/function/commands.py
@@ -16,6 +16,7 @@ from snowcli.cli.common.flags import (
     identifier_argument,
     execution_identifier_argument,
 )
+from snowcli.cli.common.project_initialisation import add_init_command
 from snowcli.cli.constants import DEPLOYMENT_STAGE, ObjectType
 from snowcli.cli.snowpark.common import (
     remove_parameter_names,
@@ -28,7 +29,6 @@ from snowcli.cli.snowpark_shared import (
     PackageNativeLibrariesOption,
     PyPiDownloadOption,
     snowpark_package,
-    ReturnsOption,
 )
 from snowcli.cli.stage.manager import StageManager
 from snowcli.exception import ObjectAlreadyExistsError
@@ -43,7 +43,6 @@ from snowcli.output.types import (
 from snowcli.utils import (
     prepare_app_zip,
     get_snowflake_packages,
-    create_project_template,
 )
 
 log = logging.getLogger(__name__)
@@ -83,15 +82,7 @@ ReplaceOption = typer.Option(
 )
 
 
-@app.command("init")
-@global_options
-@with_output
-def function_init(**options):
-    """
-    Initializes this directory with a sample set of files for creating a function.
-    """
-    create_project_template("default_function")
-    return MessageResult("Done")
+add_init_command(app, project_type="functions", template="default_function")
 
 
 @app.command("deploy")